requirements.txt
requests==2.32.4
paho-mqtt==2.1.0
pyyaml==6.0.2
build==1.2.2
setuptools>=80.9.0
